model_api/api/requirements.txt
boto3
pillow
fastapi
uvicorn
opencv-python